use baseAddress when wallet has a stake key hash
A wallet created using a mnemonic (for example, from Eternl) contains a stake key hash. This means that the addresses generated by this wallet should be longer and the base address which includes the stake key hash should be used. The enterprise address was used internally which returns shorter and a different address.

With this fix, the base address is used when the wallet has a stake key hash. This is the case when the wallet is created with a mnemonic or with a paymentSkey and stakeSkey (note: type ‘root’ not tested).

After this fix, I'm able to use both a mnemonic-created wallet and a wallet created with cardano-cli with and without stake key hash.